+++ NEWS.md
@@ -1,5 +1,7 @@
 # RPhosFate v0.11.0.9000
 
+* Renamed `"inChannelRetention"` output of `calibrationQuality()` to `"inChannelRetentionRatio"`
+* `RPhosFate` requires R ≥ 3.5.0 now
 * Slightly improved documentation
 
 # RPhosFate v0.11.0